requirements.txt
# These libraries are used ONLY for the Streamlit app.
# This is to make the Docker image as small as possible and reduce security issues.

plotly==5.17.0
streamlit==1.28.2

# Google Cloud
firebase-admin==6.2.0
google-cloud-core==2.3.3
google-cloud-bigquery==3.11.4
google-cloud-secret-manager==2.16.4